Update tiny-keccak-wrapper derivation in flake.nix
The previous derivation was using `mkDerivation`, which
(I think) was causing impurity in the build when `cargo`
invoked. This required the build to be passed
`--option sandbox false`, which can break reproducibility
and create security issues.

This update builds the `tiny-keccak-wrapper` package
"the nix way", and passes in the hash in order to ensure
reproducability.

It also adds `clang` to the build inputs to resolve
a build issue.